pyproject.toml
[tool.poetry]
name = "elenis-dirt"
version = "0.1.0"
description = ""
authors = ["Eleni-Maria <egoranites@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.10"
whitenoise = "^4.1"
gunicorn = "^20.0"
django-heroku = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
